Fix engine parts not always being added
This uses the FacepunchBehaviour Invoke method since that is the same method the game uses to delay the module entity creation. Previously, using Oxide methods like NextTick() and timer.Once() would not always wait enough time for the modules to spawn because the invoke queue is processed independent of real time.